nspawn: Do not `setsid` or surrender controlling terminal with `--console=pipe`
As far as I understand, both `setsid` calls, and the `TIOCNOTTY` ioctl were introduced in support of the "nspawn allocates a PTY" use-case.

- https://github.com/systemd/systemd/commit/a258bf2648fa5e3005b96203c60b9751807bacf0
- https://github.com/systemd/systemd/pull/13572

However, in the `--console=pipe` use-case, this results in a container with no controlling terminal, which breaks various tools. An in interactive use, it results in `Ctrl-C` in the `systemd-nspawn` container killing the entire container, which is very annoying for shell users.

Before fix:

```
$ sudo systemd-nspawn --console=pipe --as-pid2 --directory=my_chroot
-bash: cannot set terminal process group (-1): Inappropriate ioctl for device
-bash: no job control in this shell
-bash-4.4# cut -d\  -f 7 /proc/$$/stat
0
-bash-4.4# <Ctrl-C here>Container mychroot terminated by signal KILL.
<Afterwards, I need to `reset` in my outer shell to get it to be usable.>
```

After fix:

```
$ sudo ./build/systemd-nspawn --console=pipe --as-pid2 --directory=my_chroot
-bash-4.4# cut -d\  -f 7 /proc/$$/stat
34817
-bash-4.4# ^C
-bash-4.4# ^C
-bash-4.4# exit
logout
```