resolved: access control feature
Implement simple access control for systemd-resolved. This can be used to
allow-list or deny-list domains which a service can use. The feature is limited
to bus clients (such as nss-resolve).

Example directive for systemd-timesyncd.service:
[Service]
DNSAllowedDomains=*.ntp.org

Most system services have no need for DNS, so this can be enforced with:
[Service]
DNSDeniedDomains=*