resolved: access control feature
Implement simple access control for systemd-resolved. This can be used to
allow-list or deny-list domains which a service can use. The feature is limited
to bus clients (such as nss-resolve). The filtering can be used to check that
services access only approved domains (don't phone home), a minor hardening
option. A service could use DoH and direct DNS protocols to circumvent the
filtering, so those need to be blocked with other means (firewalls, SELinux).

Example directive for systemd-timesyncd.service:
[Service]
DNSAllowedDomains=*.ntp.org

Most system services have no need for DNS, so this can be enforced with:
[Service]
DNSDeniedDomains=*