Add ExtensionImages directive to form overlays
Add support for overlaying the image for services on top of their
root fs.

The overlay is not done at the root, but instead if individual
files or directories are overridden, they are bind mounted
individually.
Overlayfs is used if a directory exists both in the root and
in the image, with different content, and it is set up
read-only unconditionally.
If /usr and/or /opt are empty and not writable in the root, a tmpfs
is mounted on top.