conf-parser: configuration file downstream drop-ins should never override the main user configuration file
This patch reworks the way we parse configuration files to make sure that any
drop-in shipped by downstream (ie located in /usr/lib) never overrides user's
settings defined in the main configuration file, which is located in a "higher"
directory (eg located in /etc).

Previously all drop-ins, including those shipped by downstream, could override
the main configuration file because the main file was parsed first.

The only way to make this logic usable was by teaching users to put all theirs
settings in drop-ins only and to never use the main configuration file. However
customizing the defaults through the main file is something very well
established hence this is something nearly impossible to rely on.

Therefore the configuration drop-ins were unsable for downstream which was
unfortunate since a) our own documentation describes it as the official way for
vendors to override upstream defaults b) the only possibility left for
downstream to change upstream defaults was to patch systemd c) drop-ins allow
different flavors of a distro to use a common set of settings/drop-ins while a
specific flavor can customize the defaults further by shipping extra drop-ins.

This patch modifies our usual sorting rules to make sure that drop-ins,
regardless of their priority, located in a directory with a lowest precedence
than the directory containing the main file are always parsed before the main
file. Then the main file is parsed followed by all remaining drop-ins (located
in the same or higher directory).

This is indeed a backward incompatible change but given the fact that
downstream couldn't rely on drop-ins (because the old behavior was broken) we
can assume that no vendor was intentionnally overriding sysadmin's
customization.

Fixes: #2121