shutdown: support general timespan syntax
It can be useful to schedule a reboot which is not immediate, but where 1
minute granularity is too coarse. Our concrete use case is to initiate a reboot
in case not all critical services come up properly, with logic in the
bootloader to (eventually, after 3 attempts) falling back to another slot - we
don't want to wait +3 minutes before finally coming up, but OTOH, for debugging
and development, we do want a developer to be able to log in at the serial
console and do a 'shutdown -c' and then poke around to figure out what went
wrong.

It's also more consistent for shutdown to accept a general timespan
specification (though of course with minutes as default unit).