mount: compare mountinfo entries against current list of mounts to no reprocess a known mount continuously
With current code, every time /proc/1/mountinfo is processed, the device
and mount units for all entries is updated.
On heavily used systems, such as OpenShift nodes, we see a huge amount
of mount notifications, typically because new containers are spawned
regularly. This creates a huge load on systemd.

The optimization consists is checking whether a mount entry has already
been processed in the past, and preventing from continuously updating
the device unit and mount unit, saving CPU cycles.