logind: assume we are still docked for 5s after external connection removed
Hopefully fixes #16045 and https://bugzilla.redhat.com/show_bug.cgi?id=1840021.

When the WM is being started, reprobing of the hardware causes the external
monitor(s) to temporarily disappear. We are too quick and enter suspend when
docked. Let's wait for 5s for the external monitors to reappear before taking
action.

The timeout is set as 5s fully arbitrarily. I think this should be enough, but if
it turns out that some hardware needs more than that, we could increase it.