mount: only call device_found_node() if mount unit what has changed
The check for mount unit fields not having changed that's done in
update_parameters_proc_self_mountinfo() doesn't identify if only the
mount unit "what" parameter has changed.

But this parameter results in a function call to device_found_node()
that's only needed if the "what" parameter has changed.

So change update_parameters_proc_self_mountinfo() to return a specific
flag that indicates it has changed and arrange for device_found_node()
to be called only if the mount unit "what" parameter has changed.