CD cannot be mounted automatically (The cd was burned twice in Windows)
When udev calls the blkid_do_safeprobe function, if you assign a value to the pr->offset attribute, it will cause a disc (udf) that has been burned more than twice under windows to fail to mount and display. At first I thought this was a problem with the blkid code of util linux, and communicated with the upstream developers of blkid. For a description of the issue, see karelzak/util-linux#1161. Finally found that this was a problem with udev, and modified its code. Regarding the normal call of the blkid_do_safeprobe function, you can refer to the util-linux code: https://github.com/karelzak/util-linux/blob/master/libblkid/src/verify.c (The blkid_do_safeprobe function is called here because the blkid command is executed in the terminal, such as "blkid /dev/sr0").
Finally,@ Karelzak told me that there is no need to modify the code, just modify the rules file of udev, and my test was successful.