units: Default user services to OOMScoreAdjust=500
By setting a OOMScoreAdjust=500 on the user@.service we change the
default value for all contained user services. Doing so means that
important services that should not be killed (e.g. display server) are
able to decrease their score in relation to other applications be
reverting it to zero.

This results in disadvantaging user services against system services by
default. Doing so should not generally be an issue. It is reasonable to
do so in server scenarios and in workstation scenarios most of the
resources will be consumed by user processes so that the relative score
will be unchanged.